Move getShortDescription to utils
